Start converting data module;
It's rough.

+++ src/data/soundData.h
@@ -1,3 +1,4 @@
+#include "util.h"
 #include "data/blob.h"
 
 #pragma once
@@ -6,18 +7,18 @@ struct AudioStream;
 
 typedef struct SoundData {
   Blob blob;
-  int channelCount;
-  int sampleRate;
-  int samples;
-  int bitDepth;
+  u32 channelCount;
+  u32 sampleRate;
+  usize samples;
+  u32 bitDepth;
 } SoundData;
 
-SoundData* lovrSoundDataInit(SoundData* soundData, int samples, int sampleRate, int bitDepth, int channels);
+SoundData* lovrSoundDataInit(SoundData* soundData, usize samples, u32 sampleRate, u32 bitDepth, u32 channels);
 SoundData* lovrSoundDataInitFromAudioStream(SoundData* soundData, struct AudioStream* audioStream);
 SoundData* lovrSoundDataInitFromBlob(SoundData* soundData, Blob* blob);
 #define lovrSoundDataCreate(...) lovrSoundDataInit(lovrAlloc(SoundData), __VA_ARGS__)
 #define lovrSoundDataCreateFromAudioStream(...) lovrSoundDataInitFromAudioStream(lovrAlloc(SoundData), __VA_ARGS__)
 #define lovrSoundDataCreateFromBlob(...) lovrSoundDataInitFromBlob(lovrAlloc(SoundData), __VA_ARGS__)
-float lovrSoundDataGetSample(SoundData* soundData, int index);
-void lovrSoundDataSetSample(SoundData* soundData, int index, float value);
+f32 lovrSoundDataGetSample(SoundData* soundData, usize index);
+void lovrSoundDataSetSample(SoundData* soundData, usize index, f32 value);
 void lovrSoundDataDestroy(void* ref);
